refactor: change names of epic mode files
