+++ @remirror/extension-epic-mode/src/index.ts
@@ -1,3 +1,3 @@
-export { COLORS, defaultEffect, heartEffect, spawningEffect } from './effects';
-export { EpicModeExtension } from './epic-mode.extension';
-export * from './types';
+export { COLORS, defaultEffect, heartEffect, spawningEffect } from './epic-mode-effects';
+export { EpicModeExtension } from './epic-mode-extension';
+export * from './epic-mode-types';
